Move supixel logic into its own file, standardize constant for empty space
