+++ terminedia/__init__.py
@@ -9,9 +9,10 @@ non-blocking keyboard reading.
 from terminedia.keyboard import realtime_keyb, inkey, pause, KeyCodes
 from terminedia.utils import Rect, V2, create_transformer
 from terminedia.terminal import ScreenCommands, JournalingScreenCommands
-from terminedia.values import BlockChars, Directions, Effects, DEFAULT_BG, DEFAULT_FG, CONTEXT_COLORS, TRANSPARENT, NOP
+from terminedia.values import Directions, Effects, DEFAULT_BG, DEFAULT_FG, CONTEXT_COLORS, TRANSPARENT, NOP
 from terminedia.image import shape, ValueShape, ImageShape, PalettedShape, FullShape
 from terminedia.screen import Screen, Context
+from terminedia.subpixels import BlockChars
 from terminedia.text import render
 
 __version__ = "0.3.dev0"
